Implement an explict workaround for shadowed fields
As Python dictionaries can disambiguate between `o["get"]` and `o.get`
this version of the utility provides a way to ensure operations
are performed at the underlying map level instead of as field,
guaranteeing successful operations where the dictionary is meant
to be handled in a Pythonic way, as opposite of being handled as
regular object literal.

This new ability should allow anyone to write super robust code
without worrying of possible side-effects happening in 3rd party code.